Moving expense options dialog to be a part of the Stats_Window component because that is the component that is controlling whether or not we show the application options dialog
