+++ src/components/Stats_Window.js
@@ -4,11 +4,26 @@
  * when we size down.
  */
 
+ // Prop Type Checks on Render
+ import { PropTypes } from "prop-types";
+
+ import * as readOnlyApplicationData from "../../data/read_only_application_data";
+
 // Material UI Styles Import
 import { withStyles } from '@material-ui/core/styles';
+import { emphasize } from '@material-ui/core/styles/colorManipulator';
 
 // Material UI Components
 import Typography from '@material-ui/core/Typography';
+
+// Custom Components
+import { Options_Dialog } from './Options_Dialog';
+
+// Actions
+import {
+	closeAppOptionsDialog,
+	editApplicationOption
+} from "../actions";
 
 // Component Stylings
 const styles = theme => ({
@@ -18,6 +33,11 @@ const styles = theme => ({
 		marginRight: 0,
 		borderRadius : theme.shape.borderRadius,
 		background: theme.palette.primary.light,
+		cursor : 'pointer',
+		transition : theme.transitions.create([ 'background' ]),
+		'&:hover' : {
+			background : emphasize( theme.palette.primary.light, .30 )
+		}
 	},
 	flex : {
 		display: 'flex'
@@ -39,18 +59,31 @@ const styles = theme => ({
 	},
 });
 
-const Stats_Window = ( props ) => {
+const Stats_Window = ( props, { store } ) => {
 
 	const {
 		classes,
 		expenses,
 		expenseGroups,
 		totalCost,
-		applicationCostUnitOfMeasurement
+		applicationCostUnitOfMeasurement,
+		onClick
 	} = props;
 
+	const {
+		application_options
+	} = store.getState();
+
+	const updateApplicationOptions = ( event ) => {
+		store.dispatch(editApplicationOption({
+			[event.target.name] : event.target.value
+		}));
+	}
+
+	const { dialog_open, ...optionsValues } = application_options;
+
 	return (
-	<div className={`${classes.root} ${classes.flex}`}>
+	<div onClick={onClick} className={`${classes.root} ${classes.flex}`}>
 		<div className={`${classes.flex} ${classes.vert}`}>
 			<Typography variant="body1" color="inherit" className={classes.statsItemHeader}>Groups</Typography>
 			<Typography variant="body2" color="inherit" className={classes.statsItemValue}>{expenseGroups}</Typography>
@@ -63,7 +96,20 @@ const Stats_Window = ( props ) => {
 			<Typography variant="body1" color="inherit" className={classes.statsItemHeader}>Total Cost</Typography>
 			<Typography variant="body2" color="inherit" className={classes.statsItemValue}>${totalCost}/{applicationCostUnitOfMeasurement}</Typography>
 		</div>
+		<Options_Dialog
+			open={dialog_open}
+			title="Application"
+			options_values={optionsValues}
+			onClose={() => store.dispatch(closeAppOptionsDialog())}
+			onChange={updateApplicationOptions}
+			options_values_labels={readOnlyApplicationData["application_options_labels"]}
+			options_values_list={readOnlyApplicationData['application_options']}
+		/>
 	</div>
 )}
+
+Stats_Window.contextTypes = {
+	store : PropTypes.object.isRequired
+};
 
 export default withStyles(styles)(Stats_Window);
